Break out tests for version utils
